project/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
colorama==0.4.6
decorator==5.1.1
Django==5.0.3
djangorestframework==3.15.1
executing==2.0.1
ipython==8.22.2
jedi==0.19.1
matplotlib-inline==0.1.6
parso==0.8.3